Engine: Reenabled "Cannot use destination as argument to StrFormat" test in my_sprintf.

+++ Engine/ac/string.cpp
@@ -387,7 +387,6 @@ void my_strncpy(char *dest, const char *src, int len) {
         strcpy(dest, src);
 }
 
-// 64 bit: Not sure if this function is 64 bit ready
 // Custom printf, needed because floats are pushed as 8 bytes
 void my_sprintf(char *buffer, const char *fmt, va_list ap) {
     int bufidx = 0;
@@ -447,7 +446,7 @@ void my_sprintf(char *buffer, const char *fmt, va_list ap) {
         strncpy(fmtstring, curptr, (endptr - curptr));
         fmtstring[endptr - curptr] = 0;
 
-        unsigned long theArg = va_arg(ap, unsigned long);
+        intptr_t theArg = va_arg(ap, intptr_t);
 
         // use sprintf to parse the actual %02d type thing
         if (endptr[-1] == 'f') {
@@ -456,10 +455,8 @@ void my_sprintf(char *buffer, const char *fmt, va_list ap) {
             memcpy(&floatArg, &theArg, sizeof(float));
             sprintf(spfbuffer, fmtstring, floatArg);
         }
-/*      64 bit: Not compatible
-        else if ((theArg == (int)buffer) && (endptr[-1] == 's'))
+        else if ((theArg == (intptr_t)buffer) && (endptr[-1] == 's'))
             quit("Cannot use destination as argument to StrFormat");
-*/
         else if ((theArg < 0x10000) && (endptr[-1] == 's'))
             quit("!One of the string arguments supplied was not a string");
         else if (endptr[-1] == 's')
